feat: show adapted paper details card, when clicking a paper in the 'Papers' list (= stage entry)
